requirements.txt
aiofiles==23.2.1
amqp==5.2.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
awscli==1.29.27
billiard==4.2.0
boto3==1.28.27
botocore==1.31.27
celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
decorator==4.4.2
dnspython==2.4.2
docutils==0.16
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.101.1
googlemaps==4.10.0
h11==0.14.0
httptools==0.6.0
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.5
loguru==0.7.0
MarkupSafe==2.1.3
motor==3.2.0
moviepy==1.0.3
numpy==1.26.4
orjson==3.9.4
packaging==23.1
passlib==1.7.4
pillow==10.2.0
pluggy==1.2.0
proglog==0.1.10
prompt-toolkit==3.0.43
pyasn1==0.5.0
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.4.1
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.4
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
vine==5.1.0
watchfiles==0.19.0
wcwidth==0.2.13
websockets==11.0.3
